Les 23 is klaar vandaag ga ik tot les 26 afmaken en hoop in 45 minuten of 1 uurtje af te maken omdat de lessen heel wat korter duren

+++ app/Http/Controllers/JobController.php
@@ -1,7 +1,12 @@
 <?php
 namespace App\Http\Controllers;
 use App\Models\Job;
-use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+
+
+
 class JobController extends Controller
 {
     public function index()
@@ -32,13 +37,19 @@ class JobController extends Controller
         ]);
         return redirect('/jobs');
     }
+
+
     public function edit(Job $job)
     {
         return view('jobs.edit', ['job' => $job]);
     }
+
+
+
     public function update(Job $job)
     {
-        // authorize
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
@@ -51,7 +62,8 @@ class JobController extends Controller
     }
     public function destroy(Job $job)
     {
-        // authorize
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
         return redirect('/jobs');
     }
